Pull more code out of upvoter.ts into common modules
This will allow more sharing with opener_for_asana and perhaps future
projects.
